refactor diagnostic collection code to its own module
